requirements.txt
# Core dependencies
asyncio>=3.4.3
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=11.0.0

# Audio processing
PyAudio>=0.2.13
numpy>=1.22.0
scipy>=1.8.0

# API and networking
aiohttp>=3.8.5
requests>=2.28.0

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6

# Logging and error handling
structlog>=23.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0